fix: Authディレクトリを修正
