+++ src/components/features/Auth/index.tsx
@@ -1,20 +1,2 @@
-import { FC } from "react";
-
-import Loading from "@/components/common/Loading";
-import {
-  useAccessControl,
-  useAuth,
-} from "@/components/features/Auth/Auth.hooks";
-
-type Props = {
-  children: JSX.Element;
-};
-
-const Auth: FC<Props> = ({ children }) => {
-  const isLoading = useAuth();
-  useAccessControl();
-
-  return isLoading ? <Loading /> : children;
-};
-
-export default Auth;
+export { default } from "./Auth";
+export { useAuth, useAccessControl } from "./Auth.hooks";
